% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{base2kmer}
\alias{base2kmer}
\title{Converts a base representation (base 4 default) to the kmer}
\usage{
base2kmer(number, alph, k, base = 4)
}
\arguments{
\item{number}{A vector of integers representing a number in base 4 (default)
which correspond to a kmer.}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{k}{The length of kmers we are assessing. Used as a check.}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
A kmer, corresponding to the number in base 4(default)
}
\description{
Returns the kmer denoted by the integer
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

