% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{base5to10}
\alias{base5to10}
\title{Converts a base representation (base 4 default) to normal base 10}
\usage{
base5to10(number, k, base = 5)
}
\arguments{
\item{number}{A vector of integers representing a number in base 4 (default)
which correspond to a kmer.}

\item{k}{The length of kmers we are assessing. Used as a check.}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
An integer, corresponding to the number in base 4(default)
}
\description{
Returns an integer in base 10, calculated from the base 4 (default) 
representation
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

