% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{build_alphabet}
\alias{build_alphabet}
\title{Sets up the alphabet for indexing nucleotides}
\usage{
build_alphabet(int = c(0, 1, 2, 3, 4), let = c("G", "A", "T", "C", "M"))
}
\arguments{
\item{int}{A vector of the integers to be used, in order. Default
settings are recommended.}

\item{let}{A vector of the letters corresponding to those integers. Default
settings are recommended.}
}
\value{
Returns  a dataframe with each row being a nucleotide and an integer
rows denote equivalence betwwen the nucleotide and the integer
}
\description{
Returns a dataframe with each row being a nucleotide and an integer
}
\examples{
alph <- kmermods::build_alphabet()
int <- alph$int[alph$let=="G"]
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

