% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{convert10to5}
\alias{convert10to5}
\title{Converts a normal (base 10) integer to base representation (base 4 default)}
\usage{
convert10to5(number_b10, k, base = 5)
}
\arguments{
\item{number_b10}{An integer in base 10, representing a kmer}

\item{k}{The length of kmers we are assessing. Used as a check.}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
A vector of integers representing a number in base 4 (default)
}
\description{
Returns a base 4 (default) number as a vector calculated from the  base 10 
equivalent
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

