% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert10tobase_c}
\alias{convert10tobase_c}
\title{Converts a normal (base 10) integer to base representation (base 4 default)}
\usage{
convert10tobase_c(number_b10, k, base = 4L)
}
\arguments{
\item{number_b10}{An integer in base 10, representing a kmer}

\item{k}{The length of kmers we are assessing. Used as a check.}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
A vector of integers representing a number in base 4 (default)
}
\description{
Returns a base 4 (default) number as a vector calculated from the  base 10 
equivalent. Same as convert10to5_c but slightly faster
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

