% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{convert10tokmer}
\alias{convert10tokmer}
\title{Converts a normal (base 10) integer to the equivalent kmer}
\usage{
convert10tokmer(number_b10, alph, k, base = 5)
}
\arguments{
\item{number_b10}{An integer in base 10, representing a kmer}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{k}{The length of kmers we are assessing. Used as a check.}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
A kmer corresponding to the integer representation
}
\description{
Returns the kmer that corresponds to the integer input
}
\examples{
alph <- kmermods::build_alphabet()
convert10tokmer(112, alph, 4, base = 4)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

