% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmermods.R
\docType{data}
\name{dna_input}
\alias{dna_input}
\title{Part of a DNA sequence}
\format{A DNAStringSet obect with 1000 bp length}
\usage{
data(dna_input)
}
\description{
A DNAStringSet object (from the package Biostrings), containing 1000 bp of
the Mouse chromosome 1 sequence
}
\keyword{datasets}

