% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{kmer2base}
\alias{kmer2base}
\title{Converts a kmer to a number in base 4 (default)}
\usage{
kmer2base(kmer, alph, k)
}
\arguments{
\item{kmer}{A character vector of nucleotides to be converted the 
representation}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{k}{The length of kmers we are assessing. Used as a check.}
}
\value{
A vector of integers, representing a number in base 4(default)
}
\description{
Returns a vector of integers, which correspond to a single number in base 4
(default) or whatever is defined in build_alphabet()
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

