% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cpp_callers.R
\name{kmer_counter_c}
\alias{kmer_counter_c}
\title{Represents the kmers present in the input with their integer values, 
as a vector (with C++ implementations)}
\usage{
kmer_counter_c(dna, alph, k)
}
\arguments{
\item{dna}{The DNA sequence, as a character}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{k}{The length of kmers we are assessing. Used as a check.}
}
\value{
An integer vector corresponding to the kmers at each starting point
in the DNA
}
\description{
This runs over the input calculating the integer value for every kmer that 
begins at that location. It returns the resutls as a vector.
}
\examples{
alph <- c('G','A','T','C')
kmer_counter_c('GATCATCAT', alph, 4)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

