% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{kmer_counter_para}
\alias{kmer_counter_para}
\title{Represents the kmers present in the input with their integer values, 
as a vector, done with a parallelised version}
\usage{
kmer_counter_para(dna_input, num_cores, alph, k)
}
\arguments{
\item{dna_input}{The DNA sequence, normall a whole chromosom as a 
DNAStringSet object (see Biostrings package)}

\item{num_cores}{The number of cores to use. The default setting is the 
maximum available}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{k}{The length of kmers we are assessing. Used as a check.}
}
\value{
An integer vector corresponding to the kmers at each starting point
in the DNA
}
\description{
This runs over the input calculating the integer value for every kmer that 
begins at that location. It returns the resutls as a vector. It acts as a
wrapper to kmer_counter, and takes in DNAStringSet objects from the Biostrings
package (available from Bioconductor)
}
\examples{
\donttest{
data(dna_input)
alph <- kmermods::build_alphabet()
kmer_counter_para(dna_input,4,alph,4)}
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

