% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cpp_callers.R
\name{kmer_counter_wrapper_c}
\alias{kmer_counter_wrapper_c}
\title{Represents the kmers present in the input with their integer values, 
as a vector (with C++ implementations)}
\usage{
kmer_counter_wrapper_c(dna_input, chunk_size, alph, k)
}
\arguments{
\item{dna_input}{The DNA sequence, as a DNAStringSet object 
(see Biostrings package)}

\item{chunk_size}{The length of DNA to send in each chunk for parallel 
processing by kmer_counter}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{k}{The length of kmers we are assessing. Used as a check.}
}
\value{
An integer vector corresponding to the kmers at each starting point
in the DNA
}
\description{
This runs over the input calculating the integer value for every kmer that 
begins at that location. It returns the resutls as a vector. It acts as a
wrapper to kmer_counter, and takes in DNAStringSet objects from the Biostrings
package (available from Bioconductor)
}
\examples{
\donttest{
data(dna_input)
alph <- c('G','A','T','C')
kmer_counter_wrapper(dna_input,100,alph,4)}
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

