% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_post.R
\name{kmer_dot_prod}
\alias{kmer_dot_prod}
\title{Computes the dot product of the parameters with the kmer counts, with or 
without it being warped}
\usage{
kmer_dot_prod(kmers, paras, warp = NA)
}
\arguments{
\item{kmers}{is a vector of integers of any length representing kmers over
a window}

\item{paras}{is a vector of length equal to the total number of kmers}

\item{warp}{is a vector of length as long as the kmer vector, with the 
multiplicative weights for how much to warp the entry}
}
\value{
An double, representing the dot product of the parameters with what 
would usually be the vector of kmer abundances (warped or not)
}
\description{
Taking in an integer representing the kmer at that site, it returns the 
integer for the (k-1)mer starting at the same place
}
\examples{
kmers <- c(0,7,89,45,75,65,22,12)
paras <- rep(1,100)
kmer_dot_prod(kmers,paras)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

