\name{kmermods-package}
\alias{kmermods-package}
\alias{kmermods}
\docType{package}
\title{
\packageTitle{kmermods}
}
\description{
\packageDescription{kmermods}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kmermods}
\packageIndices{kmermods}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{kmermods}

Maintainer: \packageMaintainer{kmermods}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
