% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmermods.R
\docType{package}
\name{kmermods}
\alias{kmermods}
\alias{kmermods-package}
\title{\code{kmermods}: Represents Kmers As Integers Along The Genome}
\description{
\code{kmermods}: Represents Kmers As Integers Along The Genome
}

