% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{kmerto10}
\alias{kmerto10}
\title{Converts a kmer to it's normal (base 10) integer representation}
\usage{
kmerto10(kmer, alph, base = 5)
}
\arguments{
\item{kmer}{A character vector of nucleotides to be converted the 
representation}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
An integer corresponding to the kmer
}
\description{
Returns the integer in base 10 that corresponds to the kmer
}
\examples{
alph <- kmermods::build_alphabet()
kmerto10('GATC', alph, base = 4)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

