% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l1_prox_op}
\alias{l1_prox_op}
\title{L1 regulatisation proximal operator}
\usage{
l1_prox_op(params, thresh)
}
\arguments{
\item{params}{is a vector of length equal to the total number of kmers, 
representing the parameters in the model}

\item{thresh}{is the threshold for the proximal operator for l1 regularised
regression}
}
\value{
A vector of parameters
}
\description{
This function computes the proximal operator for L1- regularised regression
(lasso) and returns the new vector.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

