% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{let2base}
\alias{let2base}
\title{Finds the integer that represents the nucleotide}
\usage{
let2base(letter, alph)
}
\arguments{
\item{letter}{A single nucleotide, as a character.}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}
}
\value{
An integer, or NA if the input is 'N'
}
\description{
Returns an integer for {G,A,T,G} and NA for 'N'
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

