% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{let2base_c}
\alias{let2base_c}
\title{Finds the integer that represents the nucleotide}
\usage{
let2base_c(letter, alph_vect)
}
\arguments{
\item{letter}{A single nucleotide, as a character.}

\item{alph_vect}{The alphabet we are using. A dataframe created using 
build_alphabet()}
}
\value{
An integer, or NA if the input is 'N'
}
\description{
Returns an integer for {G,A,T,G} and NA for 'N'
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

