% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_reg_wrapper}
\alias{log_reg_wrapper}
\title{Computes the gradient of the log likelihood of a logistic regression}
\usage{
log_reg_wrapper(kmers_win, params, peaks, win_size, chrom_loc, warp_ = NULL)
}
\arguments{
\item{kmers_win}{is a vector of integers of any length representing kmers in
a region}

\item{peaks}{is a matrix giving the locations of the peaks on the chromosome,
the first column is starts, second is ends, inclusive, indexed from 1}

\item{win_size}{is the length of the sliding window we are using}

\item{chrom_loc}{is the position of the first kmer along the chromosome - 
this avoids indexing errors when splitting up the data
//' @param warp is a vector of length as long as the kmer vector, with the 
multiplicative weights for how much to warp the entry}

\item{paras}{is a vector of length equal to the total number of kmers}
}
\value{
A vector, representing the gradient (direction to update the 
parameter vector)
}
\description{
This calculates the prediction for each region, in a non-sliding scheme, 
calculates the error and returns the update vector for all of the regions.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

