% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loglik_linesearch}
\alias{loglik_linesearch}
\title{Calculates the log likelihood of the data, given some parameters for a 
logistic regression}
\usage{
loglik_linesearch(kmers_win, params, grad, alphas, peaks, win_size, chrom_loc,
  warp_ = NULL)
}
\arguments{
\item{kmers_win}{is a vector of integers of any length representing kmers in
a region}

\item{grad}{is a vector of length equal to the total number of kmers, 
representing the gradient of the log likelihood at that point}

\item{alphas}{is a vector of alphas, giving us the values for the line 
search, this will test loglik at params + alpha*grad for each alpha}

\item{peaks}{is a matrix giving the locations of the peaks on the chromosome,
the first column is starts, second is ends, inclusive, indexed from 1}

\item{win_size}{is the length of the sliding window we are using}

\item{chrom_loc}{is the position of the first kmer along the chromosome - 
this avoids indexing errors when splitting up the data
//' @param warp is a vector of length as long as the kmer vector, with the 
multiplicative weights for how much to warp the entry}

\item{paras}{is a vector of length equal to the total number of kmers}
}
\value{
A double, representing the log likelihood
}
\description{
Given the kmers, set of peaks and parameters, this returns the log likelihood.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

