% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mismatch_kmers}
\alias{mismatch_kmers}
\title{Counts the number of mismatches between two kmers, when represented in 
integer format}
\usage{
mismatch_kmers(kmer_1, kmer_2, k, base = 4L)
}
\arguments{
\item{kmer_1}{An integer in base 10, representing a kmer}

\item{kmer_2}{An integer in base 10, representing a kmer}

\item{k}{The length of kmers we are assessing. Used as a check.}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
A count of the mismatches
}
\description{
Returns a count of the number of mismatches between two kmers (of equal 
length), which are represented in integer format. For instance the two kmers
AAA and ATA have one mismatch. This function bypasses the conversion to 
strings.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

