% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_peaks}
\alias{predict_peaks}
\title{Computes the predictions for the logistic regression}
\usage{
predict_peaks(kmers_win, params, peaks, win_size, chrom_loc, warp_ = NULL)
}
\arguments{
\item{kmers_win}{is a vector of integers of any length representing kmers in
a region}

\item{win_size}{is the length of the sliding window we are using}

\item{chrom_loc}{is the position of the first kmer along the chromosome - 
this avoids indexing errors when splitting up the data}

\item{paras}{is a vector of length equal to the total number of kmers}

\item{warp}{is a vector of length as long as the kmer vector, with the 
multiplicative weights for how much to warp the entry}
}
\value{
A vector of probabilities
}
\description{
This calculates the predictions at each window, returning a probability of 
seeing a peak there
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

