% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cpp_callers.R
\name{rev_comp_c}
\alias{rev_comp_c}
\title{Finds the reverse complementary kmer, in integer form, without methylation 
with Rcpp implementation}
\usage{
rev_comp_c(number_b10, alph, k, base)
}
\arguments{
\item{number_b10}{An integer in base 10, representing a kmer}

\item{alph}{The alphabet we are using. A dataframe created using 
build_alphabet()}

\item{k}{The length of kmers we are assessing. Used as a check.}

\item{base}{The base of our representation, default is 4. This corresponds to
the number of distinct nucleotides we are processing.}
}
\value{
An integer, representing the reverse complementary kmer
}
\description{
Taking in an integer in base 10, it gives the integer in base 10 that 
represents the reverse complement. NOT FULLY TESTED
}
\examples{
alph <- kmermods::build_alphabet()
rev_comp_meth(100,alph,3)
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

