% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{total_error}
\alias{total_error}
\title{Computes the total error of the predictions.}
\usage{
total_error(kmers_win, params, peaks, win_size, chrom_loc, warp_ = NULL)
}
\arguments{
\item{kmers_win}{is a vector of integers of any length representing kmers in
a region}

\item{peaks}{is a matrix giving the locations of the peaks on the chromosome,
the first column is starts, second is ends, inclusive, indexed from 1}

\item{win_size}{is the length of the sliding window we are using}

\item{chrom_loc}{is the position of the first kmer along the chromosome - 
this avoids indexing errors when splitting up the data}

\item{paras}{is a vector of length equal to the total number of kmers}

\item{warp}{is a vector of length as long as the kmer vector, with the 
multiplicative weights for how much to warp the entry}
}
\value{
The total of the absolute errors, with a breakdown
}
\description{
This calculates the total of the absolute values of the errors, given a set
of parameters and outputs, over a given region.
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

