% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_paras}
\alias{update_paras}
\title{Creates the update for the parameter vector according to the error term 
and the kmers involved}
\usage{
update_paras(kmers, update_vec, err_term, add_to_update = TRUE)
}
\arguments{
\item{kmers}{A vector of integers representing the kmers that were 
present leading to the error term}

\item{update_vec}{The parameter update vector so far}

\item{err_term}{The error term used for the update}

\item{err_term}{The error term used for the update}
}
\value{
add_to_update TRUE/FALSE, whether the calculated update vector for 
the run is added automatically to the input update vector, so that the 
function returns a running total, which is the faster option.
}
\description{
Returns a vector of length equal to the parameter vector, which is update 
vector to be added to the parameter vector
}
\author{
Tom Mayo \email{t.mayo@ed.ac.uk}
}

